requirement.txt
pyautogui
pyperclip
openai
python-dotenv
